% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seurat_pipeline.R
\name{run_seurat_pipeline}
\alias{run_seurat_pipeline}
\title{run_seurat_pipeline}
\usage{
run_seurat_pipeline(object = NULL, runSCTransform = TRUE, runHarmony = TRUE)
}
\arguments{
\item{object}{seurat object}

\item{runSCTransform}{TRUE or FALSE : whether to standardize the data with SCTransform}

\item{runHarmony}{TRUE or FALSE: whether to remove batch use Harmony}
}
\value{
a seurat object
}
\description{
run_seurat_pipeline
}
\details{
this function can simply run base seurat pipeline from Seurat object to
TSNE or UMAP. However, we do not provide data quality control services.
So please do some common quality control on the data first,such as :
nFeature RNA; percent mt; cellcycle; double cells and more.
}
\examples{
 \donttest{
   pbmc <- readRDS(system.file("data","pbmc.rda",package="scrnaVis"))
   pbmc_new <- run_seurat_pipeline(object=pbmc,runSCTransform=TRUE,runHarmony=TRUE)
 }
}
