% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaVis.R
\name{scrnaVis}
\alias{scrnaVis}
\title{scrnaVis for scRNA-seq visualized elegantly}
\usage{
scrnaVis(object = NULL, markers = NULL)
}
\arguments{
\item{object}{a seurat object which has run seurat pipeline}

\item{markers}{a gene list of cell-specific expression}
}
\value{
shiny
}
\description{
scrnaVis for scRNA-seq visualized elegantly
}
\details{
This package provides multiple ways to interactively display your single-cell data
}
\examples{
\donttest{
  pbmc=readRDS(system.file("data","pbmc.rda",package="scRANvis"))
  scrnaVis(object=pbmc,markers=c("CD3E","CD3G","CD3D"))
}
}
